Analyse the position of streaks
This programmer is to examine the possible position of streaks in 3d world frame, which are detected by WISPR_Inner onboard Parker Solar Probe (PSP). 
The 01 code is purely a geometric analysis using the concept of vanishing point (projective geometry).
The feature case, in which streaks in camera are parallel to each other or intercept at one common point, are examined. For the former, only parallel line cluster in 3d frame can be projected  to the parallel one in focal plane. For the latter, we consider two possible interpretations. One is that those lines also converge at one point in 3d frame. The other is that those lines are parallel. Both of the two can result in the 2d convergence.